Moved fileSubstituter into Command
