+++ source/Calamari.Terraform/DestroyCommand.cs
@@ -5,7 +5,7 @@ namespace Calamari.Terraform
     [Command("destroy-terraform", Description = "Destroys Terraform resources")]
     public class DestroyCommand : TerraformCommand
     {
-        public DestroyCommand(): base((fileSystem, substituter) =>  new DestroyTerraformConvention(fileSystem, substituter))
+        public DestroyCommand(): base((fileSystem) =>  new DestroyTerraformConvention(fileSystem))
         {
         }
     }
