+++ source/Calamari.Terraform/DestroyPlanCommand.cs
@@ -5,7 +5,7 @@ namespace Calamari.Terraform
     [Command("destroyplan-terraform", Description = "Plans the destruction of Terraform resources")]
     public class DestroyPlanCommand : TerraformCommand
     {
-        public DestroyPlanCommand(): base((fileSystem, substituter) =>  new DestroyPlanTerraformConvention(fileSystem, substituter))
+        public DestroyPlanCommand(): base(fileSystem =>  new DestroyPlanTerraformConvention(fileSystem))
         {
         }
     }
