+++ source/Calamari.Terraform/DestroyPlanTerraformConvention.cs
@@ -1,11 +1,10 @@
 ﻿using Calamari.Integration.FileSystem;
-using Calamari.Integration.Substitutions;
 
 namespace Calamari.Terraform
 {
     public class DestroyPlanTerraformConvention : PlanTerraformConvention
     {
-        public DestroyPlanTerraformConvention(ICalamariFileSystem fileSystem, IFileSubstituter fileSubstituter) : base(fileSystem, fileSubstituter, "-destroy")
+        public DestroyPlanTerraformConvention(ICalamariFileSystem fileSystem) : base(fileSystem, "-destroy")
         {
         }
     }
