+++ source/Calamari.Terraform/PlanCommand.cs
@@ -5,7 +5,7 @@ namespace Calamari.Terraform
     [Command("plan-terraform", Description = "Plans the creation of a Terraform deployment")]
     public class PlanCommand : TerraformCommand
     {
-        public PlanCommand(): base((fileSystem, substituter) =>  new PlanTerraformConvention(fileSystem, substituter))
+        public PlanCommand(): base(fileSystem =>  new PlanTerraformConvention(fileSystem))
         {
         }
     }
